update to match capnp schema changes and fixed compiler warnings

+++ src/core/crop.cpp
@@ -119,7 +119,7 @@ void Crop::serialize(mas::models::monica::CropState::Builder builder) const {
 	_harvestDate.serialize(builder.initHarvestDate());
   if (_isWinterCrop.isValue()) builder.initIsWinterCrop().setValue(_isWinterCrop.value());
   if (_isPerennialCrop.isValue()) builder.initIsPerennialCrop().setValue(_isPerennialCrop.value());
-  setComplexCapnpList(_cuttingDates, builder.initCuttingDates(_cuttingDates.size()));
+  setComplexCapnpList(_cuttingDates, builder.initCuttingDates((uint)_cuttingDates.size()));
   if (isValid()) _cropParams.serialize(builder.initCropParams());
 	if (_separatePerennialCropParams) _separatePerennialCropParams->serialize(builder.initPerennialCropParams());
   _residueParams.serialize(builder.initResidueParams());
